+++ src/main/java/com/BANnerIt/server/api/Auth/dto/AutoLoginRequest.java
@@ -1,3 +1,3 @@
-package com.BANnerIt.server.api.user.dto;
+package com.BANnerIt.server.api.Auth.dto;
 
 public record AutoLoginRequest(String jwt) {}
