+++ src/main/java/com/BANnerIt/server/api/Auth/dto/AutoLoginResponse.java
@@ -1,0 +1,5 @@
+package com.BANnerIt.server.api.Auth.dto;
+
+import com.BANnerIt.server.api.user.dto.UserData;
+
+public record AutoLoginResponse(String jwt, UserData userData) {}
