+++ src/main/java/com/BANnerIt/server/api/Auth/repository/RefreshTokenRepository.java
@@ -1,10 +1,12 @@
-package com.BANnerIt.server.api.user.repository;
+package com.BANnerIt.server.api.Auth.repository;
 
-import com.BANnerIt.server.global.auth.RefreshToken;
+import com.BANnerIt.server.api.Auth.domain.RefreshToken;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findByUserId(Long userId);
+    void deleteByUserId(Long userId);
+    boolean existsByUserId(Long userId);
 }
