--- src/main/java/com/BANnerIt/server/api/user/dto/AutoLoginResponse.java
@@ -1,3 +1,0 @@
-package com.BANnerIt.server.api.user.dto;
-
-public record AutoLoginResponse(String jwt, UserData userData) {}
