--- src/main/java/com/BANnerIt/server/api/user/dto/LoginResponse.java
@@ -1,6 +1,0 @@
-package com.BANnerIt.server.api.user.dto;
-
-public record LoginResponse(
-        String token,
-        String message
-) {}
